Ch15 separeted the different test
