--- chapter-perf/src/test/scala/aia/performance/DispatcherTest.scala
@@ -1,462 +1,0 @@
-package aia.performance
-
-import akka.testkit.TestProbe
-import akka.actor.{ Props, ActorSystem }
-import org.scalatest.{WordSpecLike, BeforeAndAfterAll, MustMatchers}
-import monitor._
-import akka.routing.RoundRobinPool
-import monitor.ActorStatistics
-import concurrent.duration._
-import com.typesafe.config.ConfigFactory
-
-class DispatcherTest extends WordSpecLike
-  with BeforeAndAfterAll
-  with MustMatchers {
-
-  val configuration = ConfigFactory.load("performance/dispatcher")
-  implicit val system = ActorSystem("DispatcherTest", configuration)
-
-  "System" must {
-    "fails to perform" in {
-      val nrMessages = 15000
-      //      val nrMessages = 6000
-      val nrWorkers = 100
-      val statDuration = 15000 millis //((nrMessages * 10)+1000)/4 millis
-
-      //      val statProbe = TestProbe()
-      //      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-      //        storeSummaries = statProbe.ref)),"stat")
-      val printer = system.actorOf(Props[PrintMsg].withDispatcher("my-pinned-dispatcher"), "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor)),
-        "Workers")
-
-      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      //      val statMsg = statProbe.expectMsgAllOf[Any]()
-      //      println(statMsg)
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-!!!!!!PRINT!!!!!!!! nr=15
-ENTRY: maxQueue=70, utilization 5,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235660000,1376235675000,70,5.0,4240.0,5.0,10.0)
-ENTRY: maxQueue=179, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235675000,1376235690000,179,8.0,15774.0,8.0,10.0)
-ENTRY: maxQueue=285, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235690000,1376235705000,285,8.0,29124.0,8.0,10.0)
-ENTRY: maxQueue=385, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235705000,1376235720000,385,7.0,42193.0,7.0,10.0)
-ENTRY: maxQueue=490, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235720000,1376235735000,490,8.0,55268.0,8.0,10.0)
-ENTRY: maxQueue=595, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235735000,1376235750000,595,8.0,68617.0,8.0,10.0)
-ENTRY: maxQueue=705, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235750000,1376235765000,705,8.0,81967.0,8.0,10.0)
-ENTRY: maxQueue=810, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235765000,1376235780000,810,8.0,95317.0,8.0,10.0)
-ENTRY: maxQueue=915, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235780000,1376235795000,915,8.0,108667.0,8.0,10.0)
-ENTRY: maxQueue=1013, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235795000,1376235810000,1013,7.0,121913.0,7.0,10.0)
-ENTRY: maxQueue=1121, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235810000,1376235825000,1121,7.0,134752.0,7.0,10.0)
-ENTRY: maxQueue=1225, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235825000,1376235840000,1225,7.0,147656.0,7.0,10.0)
-ENTRY: maxQueue=1325, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235840000,1376235855000,1325,8.0,160955.0,8.0,10.0)
-ENTRY: maxQueue=1430, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235855000,1376235870000,1430,8.0,174305.0,8.0,10.0)
-ENTRY: maxQueue=1540, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235870000,1376235885000,1540,8.0,187655.0,8.0,10.0)
-
-
-ENTRY: maxQueue=95, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233425000,1376233440000,95,7.0,5421.0,7.0,10.0)
-      ENTRY: maxQueue=200, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233440000,1376233455000,200,8.0,18492.0,8.0,10.0)
-      ENTRY: maxQueue=310, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233455000,1376233470000,310,8.0,32444.0,8.0,10.0)
-      ENTRY: maxQueue=425, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233470000,1376233485000,425,8.0,46394.0,8.0,10.0)
-      ENTRY: maxQueue=535, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233485000,1376233500000,535,8.0,60342.0,8.0,10.0)
-      ENTRY: maxQueue=645, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233500000,1376233515000,645,8.0,74292.0,8.0,10.0)
-      ENTRY: maxQueue=749, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233515000,1376233530000,749,7.0,87953.0,7.0,10.0)
-      ENTRY: maxQueue=860, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233530000,1376233545000,860,8.0,101611.0,8.0,10.0)
-      ENTRY: maxQueue=970, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233545000,1376233560000,970,8.0,115562.0,8.0,10.0)
-      ENTRY: maxQueue=1080, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233560000,1376233575000,1080,8.0,129513.0,8.0,10.0)
-      ENTRY: maxQueue=1195, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233575000,1376233590000,1195,8.0,143464.0,8.0,10.0)
-      ENTRY: maxQueue=1302, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233590000,1376233605000,1302,8.0,157415.0,8.0,10.0)
-      ENTRY: maxQueue=1405, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233605000,1376233620000,1405,7.0,171090.0,7.0,10.0)
-      ENTRY: maxQueue=1510, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233620000,1376233635000,1510,7.0,184437.0,7.0,10.0)
-      ENTRY: maxQueue=1625, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233635000,1376233650000,1625,8.0,198105.0,8.0,10.0)
-      ENTRY: maxQueue=1735, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233650000,1376233665000,1735,8.0,212055.0,8.0,10.0)
-      ENTRY: maxQueue=1850, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233665000,1376233680000,1850,8.0,226005.0,8.0,10.0)
-      ENTRY: maxQueue=1955, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233680000,1376233695000,1955,8.0,239954.0,8.0,10.0)
-      ENTRY: maxQueue=2065, utilization 8,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233695000,1376233710000,2065,8.0,253905.0,8.0,10.0)
-      ENTRY: maxQueue=2170, utilization 7,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376233710000,1376233725000,2170,7.0,267571.0,7.0,10.0)
-      StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$wb],1376233425000,1376233440000,1,0.0,33.0,6.0,1000.0)
-*/
-    }
-    "Workers fails to perform" in {
-      val nrMessages = 15000
-      //      val nrMessages = 6000
-      val nrWorkers = 100
-      val statDuration = 15000 millis //((nrMessages * 10)+1000)/4 millis
-
-      //      val statProbe = TestProbe()
-      //      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-      //        storeSummaries = statProbe.ref)),"stat")
-      val printer = system.actorOf(Props[PrintMsg], "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor)),
-        "Workers")
-
-      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor).withDispatcher("my-pinned-dispatcher"), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      //      val statMsg = statProbe.expectMsgAllOf[Any]()
-      //      println(statMsg)
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1376236035000,1376236050000,10,0.0,33574.0,21.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1376236035000,1376236050000,14,0.0,19902.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376236035000,1376236050000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1376236035000,1376236050000,11,0.0,58427.0,5.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1376236035000,1376236050000,9,0.0,41049.0,18.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1376236035000,1376236050000,15,0.0,47485.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1376236035000,1376236050000,20,0.0,43690.0,61.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1376236035000,1376236050000,15,0.0,40006.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1376236035000,1376236050000,10,0.0,27453.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1376236035000,1376236050000,5,0.0,67365.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1376236035000,1376236050000,15,0.0,16144.0,48.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1376236035000,1376236050000,21,0.0,24713.0,54.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1376236035000,1376236050000,17,0.0,21897.0,78.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376236035000,1376236050000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1376236035000,1376236050000,5,0.0,54695.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1376236035000,1376236050000,20,0.0,43956.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1376236035000,1376236050000,18,0.0,38310.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1376236035000,1376236050000,17,0.0,46801.0,60.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1376236035000,1376236050000,7,0.0,17620.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236035000,1376236050000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1376236035000,1376236050000,8,0.0,26196.0,18.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1376236035000,1376236050000,6,0.0,21092.0,5.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1376236035000,1376236050000,6,0.0,65492.0,38.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1376236035000,1376236050000,5,0.0,13051.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1376236035000,1376236050000,8,0.0,64623.0,51.0,1000.0))
-!!!!!!PRINT!!!!!!!! nr=15
-ENTRY: maxQueue=2, utilization 65,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235975000,1376235990000,2,65.0,0.0,65.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376235990000,1376236005000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236005000,1376236020000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236020000,1376236035000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236035000,1376236050000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236050000,1376236065000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236065000,1376236080000,1,63.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236080000,1376236095000,1,66.0,0.0,66.0,9.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236095000,1376236110000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236110000,1376236125000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236125000,1376236140000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236140000,1376236155000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236155000,1376236170000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236170000,1376236185000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376236185000,1376236200000,1,66.0,0.0,66.0,10.0)
-
-StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1376236200000,1376236215000,25,0.0,200305.0,33.0,1000.0),
-StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1376236200000,1376236215000,40,0.0,179679.0,33.0,1000.0),
-StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1376236200000,1376236215000,18,0.0,209638.0,47.0,1000.0),
-StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$sb],1376236200000,1376236215000,5,0.0,230841.0,33.0,1000.0),
-
-       */
-
-    }
-    "Use multiple threads" in {
-      val nrMessages = 15000
-      val nrWorkers = 100
-      val statDuration = 15000 millis
-
-      val printer = system.actorOf(Props[PrintMsg], "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
-            withDispatcher("myMultiThread-dispatcher")),
-        "Workers")
-
-      val firstStep = system.actorOf(
-        Props(new ProcessRequest(10 millis, workers) with MonitorActor).
-          withDispatcher("my-pinned-dispatcher"), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-!!!!!!PRINT!!!!!!!! nr=15
-ENTRY: maxQueue=2, utilization 34,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376821995000,1376822010000,2,34.0,0.0,34.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822010000,1376822025000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822025000,1376822040000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822040000,1376822055000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822055000,1376822070000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822070000,1376822085000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822085000,1376822100000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822100000,1376822115000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822115000,1376822130000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822130000,1376822145000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822145000,1376822160000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822160000,1376822175000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822175000,1376822190000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822190000,1376822205000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822205000,1376822220000,1,66.0,0.0,66.0,10.0)
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/$A],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ib],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$V],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$jb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376821995000,1376822010000,3,65.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$8],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$N],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$F],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$xb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$g],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Jb],1376821995000,1376822010000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$c],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$cb],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$4],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$9],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$y],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$D],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$0],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$s],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$O],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$w],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$K],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ob],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ab],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$nb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$vb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$q],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Db],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$5],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$m],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$i],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$U],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$E],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$S],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$lb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$eb],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$qb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$tb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$hb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Fb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$z],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$~],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$mb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$B],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$2],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$h],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$d],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Eb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$n],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$j],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$H],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$t],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$x],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$L],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$P],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$X],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$db],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$1],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$wb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$pb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$yb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Hb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$k],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376821995000,1376822010000,1,30.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$6],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$o],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$sb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$gb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$M],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$I],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$C],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Y],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$a],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$u],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ib],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$+],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Q],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Gb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$G],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$W],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$e],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$rb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$b],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ub],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376821995000,1376822010000,2,34.0,0.0,34.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$R],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ab],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$J],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Z],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$3],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Cb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$kb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$v],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$7],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$r],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$l],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$p],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Bb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$fb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$T],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$zb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$bb],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$f],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0))
-
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/$A],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ib],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$V],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$jb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376822205000,1376822220000,8,132.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$8],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$N],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$F],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$xb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$g],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Jb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$c],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$cb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$4],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$9],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$y],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$D],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$0],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$s],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$O],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$w],1376822205000,1376822220000,1,0.0,0.0,66.0,1002.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$K],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ob],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ab],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$nb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$vb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$q],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Db],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$5],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$m],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$i],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$U],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$E],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$S],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$lb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$eb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$qb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$tb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$hb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Fb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$z],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$~],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$mb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$B],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$2],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$h],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$d],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Eb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$n],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$j],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$H],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$t],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$x],1376822205000,1376822220000,1,0.0,0.0,66.0,1001.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$L],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$P],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$X],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$db],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$1],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$wb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$pb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$yb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Hb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$k],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376822205000,1376822220000,1,66.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$6],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$o],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$sb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$gb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$M],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$I],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$C],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Y],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$a],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$u],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ib],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$+],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Q],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Gb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$G],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$W],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$e],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$rb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$b],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ub],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822205000,1376822220000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$R],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ab],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$J],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Z],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$3],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1376822205000,1376822220000,1,0.0,1.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Cb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$kb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$v],1376822205000,1376822220000,1,0.0,0.0,66.0,1004.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$7],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$r],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$l],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$p],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Bb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$fb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$T],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$zb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$bb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$f],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0))
-
-       */
-
-    }
-    "Use multiple threads2" in {
-      val nrMessages = 15000
-      val nrWorkers = 100
-      val statDuration = 15000 millis
-
-      val printer = system.actorOf(Props[PrintMsg], "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
-            withDispatcher("myMultiThread-dispatcher")),
-        "Workers")
-
-      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor).withDispatcher("myMultiThread-dispatcher"), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-!!!!!!PRINT!!!!!!!! nr=15
-ENTRY: maxQueue=3, utilization 29,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822910000,1376822925000,3,29.0,0.0,29.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822925000,1376822940000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822940000,1376822955000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822955000,1376822970000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822970000,1376822985000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822985000,1376823000000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823000000,1376823015000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823015000,1376823030000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823030000,1376823045000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823045000,1376823060000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823060000,1376823075000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823075000,1376823090000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823090000,1376823105000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823105000,1376823120000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823120000,1376823135000,1,66.0,0.0,66.0,10.0)
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$wb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$l],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$e],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376822910000,1376822925000,2,54.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$qb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$h],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ib],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$b],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$fb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$sb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$jb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$f],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Cb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$rb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Gb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Hb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$vb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Eb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$gb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$i],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$kb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$c],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$eb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ob],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ib],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ub],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Db],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$j],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Bb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376822910000,1376822925000,1,25.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$g],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers],1376822910000,1376822925000,100,6.0,19.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$nb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$hb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$zb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$lb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$d],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$yb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$xb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$pb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Jb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822910000,1376822925000,3,29.0,0.0,29.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$tb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Fb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ab],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$a],1376822910000,1376822925000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$k],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$mb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0))
-
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$wb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$l],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$e],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376823105000,1376823120000,9,132.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$qb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$h],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ib],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$b],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$fb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$sb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$jb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$f],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Cb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$rb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Gb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Hb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$vb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Eb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$gb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$i],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$kb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$c],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$eb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ob],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ib],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ub],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1376823105000,1376823120000,1,0.0,0.0,66.0,1001.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Db],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$j],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Bb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376823105000,1376823120000,1,66.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$g],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$nb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$hb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$zb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$lb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$d],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$yb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$xb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$pb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Jb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823105000,1376823120000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$tb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1376823105000,1376823120000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Fb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ab],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1376823105000,1376823120000,1,0.0,0.0,66.0,1003.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$a],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1376823105000,1376823120000,1,0.0,0.0,66.0,1004.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$k],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$mb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0))
-       */
-
-    }
-    "fails to perform change throughput" in {
-      val nrMessages = 15000
-      val nrWorkers = 100
-      val statDuration = 15000 millis
-
-      val printer = system.actorOf(Props[PrintMsg].withDispatcher("my-pinned-dispatcher"), "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
-            withDispatcher("throughput-dispatcher")),
-        "Workers")
-
-      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor).withDispatcher("throughput-dispatcher"), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-      !!!!!!PRINT!!!!!!!! nr=16
-ENTRY: maxQueue=62, utilization 13,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347145000,1377347160000,62,13.0,444.0,13.0,10.0)
-ENTRY: maxQueue=162, utilization 68,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347160000,1377347175000,162,68.0,1006.0,68.0,10.0)
-ENTRY: maxQueue=22, utilization 2,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347175000,1377347190000,22,2.0,106.0,2.0,10.0)
-ENTRY: maxQueue=1000, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347190000,1377347205000,1000,66.0,13638.0,66.0,10.0)
-ENTRY: maxQueue=1076, utilization 49,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347250000,1377347265000,1076,49.0,65285.0,49.0,10.0)
-ENTRY: maxQueue=1000, utilization 17,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347265000,1377347280000,1000,17.0,62787.0,17.0,10.0)
-
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347295000,1377347310000,30,1.0,74768.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347295000,1377347310000,2,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347295000,1377347310000,12,0.0,137622.0,76.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347295000,1377347310000,31,1.0,70230.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1377347295000,1377347310000,10,0.0,42402.0,62.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347295000,1377347310000,19,1.0,122255.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347295000,1377347310000,33,0.0,32120.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1377347295000,1377347310000,9,0.0,43363.0,56.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347295000,1377347310000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347295000,1377347310000,33,0.0,35321.0,58.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1377347295000,1377347310000,33,0.0,33082.0,43.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1377347295000,1377347310000,10,0.0,30429.0,23.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1377347295000,1377347310000,7,0.0,45603.0,41.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347295000,1377347310000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347295000,1377347310000,32,1.0,65047.0,100.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347310000,1377347325000,33,0.0,44896.0,22.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1377347310000,1377347325000,10,0.0,52926.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347310000,1377347325000,2,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347310000,1377347325000,27,1.0,105218.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347310000,1377347325000,13,0.0,151399.0,82.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347310000,1377347325000,33,0.0,44092.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347310000,1377347325000,33,0.0,71917.0,96.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347310000,1377347325000,15,1.0,146960.0,96.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347310000,1377347325000,2,0.0,165312.0,10.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1377347310000,1377347325000,10,0.0,55188.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1377347310000,1377347325000,12,0.0,152782.0,74.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347310000,1377347325000,4,0.0,162482.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1377347310000,1377347325000,10,0.0,43361.0,10.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347310000,1377347325000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347310000,1377347325000,1,0.0,66208.0,3.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1377347310000,1377347325000,10,0.0,45600.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347310000,1377347325000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347310000,1377347325000,14,0.0,149057.0,89.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347310000,1377347325000,33,0.0,42644.0,7.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1377347310000,1377347325000,0,0.0,0.0,3.0,0.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347325000,1377347340000,2,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347325000,1377347340000,33,0.0,63070.0,43.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347325000,1377347340000,28,1.0,114371.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1377347325000,1377347340000,6,0.0,174969.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347325000,1377347340000,30,1.0,104912.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347325000,1377347340000,17,1.0,155794.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1377347325000,1377347340000,27,1.0,120379.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347325000,1377347340000,19,1.0,151894.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347325000,1377347340000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347325000,1377347340000,10,0.0,66206.0,62.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347325000,1377347340000,9,0.0,171353.0,56.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347325000,1377347340000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347325000,1377347340000,29,1.0,109608.0,100.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347340000,1377347355000,2,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347340000,1377347355000,33,0.0,77032.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1377347340000,1377347355000,21,1.0,161612.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347340000,1377347355000,33,0.0,75040.0,23.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347340000,1377347355000,32,1.0,106487.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1377347340000,1377347355000,33,0.0,78231.0,45.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347340000,1377347355000,33,0.0,101557.0,94.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347340000,1377347355000,10,0.0,85342.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1377347340000,1377347355000,9,0.0,88542.0,54.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347340000,1377347355000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1377347340000,1377347355000,11,0.0,183703.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347340000,1377347355000,24,1.0,147769.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347340000,1377347355000,2,0.0,95314.0,9.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1377347340000,1377347355000,10,0.0,87344.0,62.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347340000,1377347355000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347340000,1377347355000,33,0.0,76002.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1377347340000,1377347355000,1,0.0,196001.0,5.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347355000,1377347370000,2,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1377347355000,1377347370000,10,0.0,97284.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1377347355000,1377347370000,32,0.0,103918.0,76.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$fb],1377347355000,1377347370000,7,0.0,203872.0,42.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$jb],1377347355000,1377347370000,4,0.0,208038.0,21.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$gb],1377347355000,1377347370000,5,0.0,206204.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$eb],1377347355000,1377347370000,7,0.0,203862.0,42.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347355000,1377347370000,10,0.0,98482.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ib],1377347355000,1377347370000,4,0.0,207682.0,23.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1377347355000,1377347370000,10,0.0,88540.0,11.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347355000,1377347370000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1377347355000,1377347370000,26,1.0,151082.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347355000,1377347370000,32,0.0,96002.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$hb],1377347355000,1377347370000,5,0.0,206380.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347355000,1377347370000,10,0.0,95311.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347355000,1377347370000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347355000,1377347370000,10,0.0,97108.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1377347355000,1377347370000,16,1.0,187915.0,100.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347355000,1377347370000,0,0.0,0.0,2.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1377347355000,1377347370000,0,0.0,0.0,3.0,0.0))
-
-
-      */
-    }
-    "fails to perform seperate dispatchers" in {
-      val nrMessages = 15000
-      val nrWorkers = 100
-      val statDuration = 15000 millis
-
-      val printer = system.actorOf(Props[PrintMsg], "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
-            withDispatcher("my-dispatcher2")),
-        "Workers")
-
-      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-!!!!!!PRINT!!!!!!!! nr=16
-ENTRY: maxQueue=2, utilization 15,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347685000,1377347700000,2,15.0,0.0,15.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347700000,1377347715000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347715000,1377347730000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347730000,1377347745000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347745000,1377347760000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347760000,1377347775000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347775000,1377347790000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347790000,1377347805000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347805000,1377347820000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 67,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347820000,1377347835000,1,65.0,0.0,67.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347835000,1377347850000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347850000,1377347865000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347865000,1377347880000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347880000,1377347895000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347895000,1377347910000,1,66.0,0.0,66.0,10.0)
-ENTRY: maxQueue=1, utilization 52,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347910000,1377347925000,1,52.0,0.0,52.0,10.0)
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$l],1377347685000,1377347700000,1,0.0,636.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1377347685000,1377347700000,1,0.0,2777.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$e],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1377347685000,1377347700000,1,0.0,2762.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1377347685000,1377347700000,1,0.0,631.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347685000,1377347700000,3,16.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$h],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$b],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1377347685000,1377347700000,1,0.0,629.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1377347685000,1377347700000,1,0.0,2772.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1377347685000,1377347700000,1,0.0,629.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$f],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1377347685000,1377347700000,1,0.0,2760.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$i],1377347685000,1377347700000,1,0.0,657.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$c],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1377347685000,1377347700000,1,0.0,629.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1377347685000,1377347700000,1,0.0,2767.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1377347685000,1377347700000,1,0.0,2760.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$j],1377347685000,1377347700000,1,0.0,646.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347685000,1377347700000,1,1.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$g],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers],1377347685000,1377347700000,85,6.0,8.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$d],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1377347685000,1377347700000,1,0.0,2788.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347685000,1377347700000,2,15.0,0.0,15.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1377347685000,1377347700000,1,0.0,2760.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$a],1377347685000,1377347700000,1,0.0,0.0,6.0,1007.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$k],1377347685000,1377347700000,1,0.0,641.0,13.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1377347700000,1377347715000,5,0.0,11579.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1377347700000,1377347715000,3,0.0,2355.0,72.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1377347700000,1377347715000,2,0.0,1182.0,50.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1377347700000,1377347715000,6,0.0,5408.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347700000,1377347715000,5,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1377347700000,1377347715000,2,0.0,1519.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1377347700000,1377347715000,2,0.0,1256.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1377347700000,1377347715000,5,0.0,6649.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1377347700000,1377347715000,2,0.0,1263.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1377347700000,1377347715000,2,0.0,1256.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347700000,1377347715000,6,0.0,5401.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1377347700000,1377347715000,6,0.0,10558.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347700000,1377347715000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347700000,1377347715000,6,0.0,7309.0,42.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347700000,1377347715000,1,0.0,17567.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347700000,1377347715000,6,0.0,11314.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1377347700000,1377347715000,2,0.0,1246.0,50.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347700000,1377347715000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1377347700000,1377347715000,2,0.0,1191.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1377347700000,1377347715000,6,0.0,5164.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1377347700000,1377347715000,7,0.0,10114.0,42.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347700000,1377347715000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347700000,1377347715000,1,0.0,17565.0,2.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1377347715000,1377347730000,10,0.0,11656.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1377347715000,1377347730000,1,0.0,578.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347715000,1377347730000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1377347715000,1377347730000,5,0.0,21515.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347715000,1377347730000,6,0.0,24611.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347715000,1377347730000,6,0.0,22909.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347715000,1377347730000,4,0.0,3503.0,75.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347715000,1377347730000,5,0.0,26459.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1377347715000,1377347730000,8,0.0,8837.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347715000,1377347730000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347715000,1377347730000,3,0.0,29988.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347715000,1377347730000,7,0.0,6288.0,87.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347715000,1377347730000,7,0.0,18810.0,40.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347715000,1377347730000,5,0.0,26474.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347715000,1377347730000,9,0.0,7826.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347715000,1377347730000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1377347715000,1377347730000,3,0.0,1161.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1377347715000,1377347730000,8,0.0,5511.0,90.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347715000,1377347730000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347715000,1377347730000,12,0.0,13810.0,97.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1377347730000,1377347745000,9,0.0,8803.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347730000,1377347745000,2,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1377347730000,1377347745000,5,0.0,36426.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347730000,1377347745000,16,0.0,24262.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347730000,1377347745000,11,0.0,34929.0,68.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347730000,1377347745000,11,0.0,25524.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347730000,1377347745000,1,0.0,422.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347730000,1377347745000,15,0.0,26202.0,62.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347730000,1377347745000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347730000,1377347745000,6,0.0,37802.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347730000,1377347745000,15,0.0,27444.0,84.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347730000,1377347745000,7,0.0,10134.0,46.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347730000,1377347745000,12,0.0,21940.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347730000,1377347745000,15,0.0,26749.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347730000,1377347745000,7,0.0,6308.0,97.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347730000,1377347745000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347730000,1377347745000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1377347730000,1377347745000,3,0.0,44928.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347730000,1377347745000,11,0.0,10826.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347730000,1377347745000,3,0.0,44929.0,15.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347745000,1377347760000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1377347745000,1377347760000,6,0.0,55242.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347745000,1377347760000,18,0.0,21225.0,97.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347745000,1377347760000,3,0.0,59839.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347745000,1377347760000,15,0.0,31188.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347745000,1377347760000,15,0.0,31247.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1377347745000,1377347760000,5,0.0,51382.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347745000,1377347760000,19,0.0,26196.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347745000,1377347760000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347745000,1377347760000,16,0.0,37662.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347745000,1377347760000,21,0.0,25395.0,75.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347745000,1377347760000,8,0.0,7002.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347745000,1377347760000,17,0.0,27285.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347745000,1377347760000,16,0.0,39952.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347745000,1377347760000,5,0.0,56337.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347745000,1377347760000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347745000,1377347760000,6,0.0,52742.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347745000,1377347760000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1377347745000,1377347760000,10,0.0,47382.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347745000,1377347760000,9,0.0,17067.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347745000,1377347760000,13,0.0,43187.0,66.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347760000,1377347775000,3,0.0,74764.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347760000,1377347775000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1377347760000,1377347775000,16,0.0,54093.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347760000,1377347775000,5,0.0,71250.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347760000,1377347775000,1,0.0,77235.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347760000,1377347775000,17,0.0,21473.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347760000,1377347775000,5,0.0,58587.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347760000,1377347775000,20,0.0,41951.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1377347760000,1377347775000,10,0.0,64833.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347760000,1377347775000,19,0.0,28991.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347760000,1377347775000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347760000,1377347775000,20,0.0,38669.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347760000,1377347775000,19,0.0,21818.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347760000,1377347775000,11,0.0,64306.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1377347760000,1377347775000,10,0.0,64071.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347760000,1377347775000,18,0.0,29425.0,48.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347760000,1377347775000,19,0.0,38698.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347760000,1377347775000,6,0.0,69794.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347760000,1377347775000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347760000,1377347775000,15,0.0,54568.0,64.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347760000,1377347775000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347760000,1377347775000,5,0.0,69293.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347760000,1377347775000,13,0.0,18258.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347760000,1377347775000,22,0.0,44880.0,60.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347775000,1377347790000,8,0.0,78770.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347775000,1377347790000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1377347775000,1377347790000,20,0.0,53563.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347775000,1377347790000,5,0.0,79233.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347775000,1377347790000,5,0.0,85150.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347775000,1377347790000,10,0.0,79713.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347775000,1377347790000,5,0.0,75982.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347775000,1377347790000,5,0.0,81160.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347775000,1377347790000,15,0.0,69553.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347775000,1377347790000,1,0.0,92099.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347775000,1377347790000,25,0.0,52385.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1377347775000,1377347790000,5,0.0,86126.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1377347775000,1377347790000,15,0.0,66347.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347775000,1377347790000,3,0.0,89659.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347775000,1377347790000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347775000,1377347790000,21,0.0,62376.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347775000,1377347790000,18,0.0,33954.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347775000,1377347790000,17,0.0,63522.0,40.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347775000,1377347790000,1,0.0,92103.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347775000,1377347790000,20,0.0,28267.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347775000,1377347790000,15,0.0,67308.0,64.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347775000,1377347790000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347775000,1377347790000,25,0.0,56602.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347775000,1377347790000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347775000,1377347790000,6,0.0,84295.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347775000,1377347790000,10,0.0,78763.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347775000,1377347790000,30,0.0,44267.0,57.0,1000.0))
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347790000,1377347805000,15,0.0,85433.0,47.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347790000,1377347805000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347790000,1377347805000,15,0.0,86451.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347790000,1377347805000,10,0.0,93639.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347790000,1377347805000,18,0.0,82240.0,52.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1377347790000,1377347805000,5,0.0,96066.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347790000,1377347805000,7,0.0,98460.0,9.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347790000,1377347805000,11,0.0,91791.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347790000,1377347805000,20,0.0,73773.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347790000,1377347805000,5,0.0,90846.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347790000,1377347805000,25,0.0,56961.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347790000,1377347805000,10,0.0,88573.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347790000,1377347805000,19,0.0,48207.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347790000,1377347805000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347790000,1377347805000,26,0.0,62883.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1377347790000,1377347805000,1,0.0,107043.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347790000,1377347805000,25,0.0,69774.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347790000,1377347805000,5,0.0,90850.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1377347790000,1377347805000,13,0.0,89819.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347790000,1377347805000,20,0.0,36544.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347790000,1377347805000,20,0.0,72808.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347790000,1377347805000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347790000,1377347805000,30,0.0,61886.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347790000,1377347805000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347790000,1377347805000,10,0.0,83435.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347790000,1377347805000,15,0.0,85251.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1377347790000,1377347805000,1,0.0,107058.0,2.0,1000.0))
-
-
-      */
-    }
-    "fails to perform threadPool" in {
-      val nrMessages = 15000
-      val nrWorkers = 100
-      val statDuration = 15000 millis
-
-      val printer = system.actorOf(Props[PrintMsg], "Printer")
-      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
-        storeSummaries = printer)), "stat")
-      system.eventStream.subscribe(
-        stat,
-        classOf[ActorStatistics])
-      system.eventStream.subscribe(
-        stat,
-        classOf[MailboxStatistics])
-
-      val end = TestProbe()
-
-      val workers = system.actorOf(
-        RoundRobinPool(nrWorkers).props(
-          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
-            withDispatcher("my-thread-dispatcher")),
-        "Workers")
-
-      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor), "Entry")
-
-      for (i <- 0 until nrMessages) {
-        firstStep ! new SystemMessage()
-        Thread.sleep(15)
-      }
-      Thread.sleep(10000)
-      printer ! "print"
-      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
-
-      system.stop(firstStep)
-      system.stop(workers)
-      system.stop(stat)
-      system.stop(printer)
-      /*
-
-Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1378984500000,1378984515000,15,0.0,174146.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1378984500000,1378984515000,15,0.0,163621.0,14.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1378984500000,1378984515000,15,0.0,169265.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1378984500000,1378984515000,15,0.0,169252.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1378984500000,1378984515000,15,0.0,168136.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1378984500000,1378984515000,5,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1378984500000,1378984515000,13,0.0,179513.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1378984500000,1378984515000,15,0.0,174114.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1378984500000,1378984515000,15,0.0,168128.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1378984500000,1378984515000,15,0.0,163606.0,14.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1378984500000,1378984515000,15,0.0,168124.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1378984500000,1378984515000,15,0.0,169243.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1378984500000,1378984515000,13,0.0,179529.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1378984500000,1378984515000,13,0.0,179498.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1378984500000,1378984515000,13,0.0,179526.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1378984500000,1378984515000,15,0.0,169237.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1378984500000,1378984515000,15,0.0,168121.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1378984500000,1378984515000,15,0.0,163603.0,14.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1378984500000,1378984515000,15,0.0,169250.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1378984500000,1378984515000,15,0.0,174145.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1378984500000,1378984515000,13,0.0,179504.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1378984500000,1378984515000,15,0.0,174133.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1378984500000,1378984515000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1378984500000,1378984515000,13,0.0,179495.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1378984500000,1378984515000,13,0.0,179527.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1378984500000,1378984515000,15,0.0,169234.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1378984500000,1378984515000,15,0.0,174132.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1378984500000,1378984515000,13,0.0,179511.0,19.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1378984500000,1378984515000,15,0.0,174130.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1378984500000,1378984515000,15,0.0,163635.0,14.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1378984500000,1378984515000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1378984500000,1378984515000,15,0.0,169268.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1378984500000,1378984515000,15,0.0,169266.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1378984500000,1378984515000,15,0.0,174138.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1378984500000,1378984515000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1378984500000,1378984515000,15,0.0,174117.0,33.0,1000.0))
-
-      */
-    }
-
-  }
-}
-
