+++ chapter-perf/src/test/scala/aia/performance/dispatcher/DispatcherSeperateTest.scala
@@ -1,0 +1,91 @@
+package aia.performance.dispatcher
+
+import akka.testkit.TestProbe
+import akka.actor.{Props, ActorSystem}
+import org.scalatest.{WordSpecLike, BeforeAndAfterAll, MustMatchers}
+import akka.routing.RoundRobinPool
+import aia.performance.monitor.ActorStatistics
+import com.typesafe.config.ConfigFactory
+import aia.performance.{SystemMessage, ProcessRequest, PrintMsg}
+import aia.performance.monitor.{MonitorActor, MailboxStatistics, ActorStatistics, MonitorStatisticsActor}
+import concurrent.duration._
+
+class DispatcherSeperateTest extends WordSpecLike
+  with BeforeAndAfterAll
+  with MustMatchers {
+
+  val configuration = ConfigFactory.load("performance/dispatcher")
+  implicit val system = ActorSystem("DispatcherTest", configuration)
+
+  "System" must {
+     "fails to perform seperate dispatchers" in {
+       //val nrMessages = 15000
+       val nrMessages = 6000
+       val nrWorkers = 100
+      val statDuration = 15000 millis
+
+      val printer = system.actorOf(Props[PrintMsg], "Printer-dispatcher")
+      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
+        storeSummaries = printer)), "stat-dispatcher")
+      system.eventStream.subscribe(
+        stat,
+        classOf[ActorStatistics])
+      system.eventStream.subscribe(
+        stat,
+        classOf[MailboxStatistics])
+
+      val end = TestProbe()
+
+      val workers = system.actorOf(
+        RoundRobinPool(nrWorkers).props(
+          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
+            withDispatcher("my-dispatcher2")),
+        "Workers")
+
+      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor), "Entry")
+
+      for (i <- 0 until nrMessages) {
+        firstStep ! new SystemMessage()
+        Thread.sleep(15)
+      }
+      Thread.sleep(10000)
+      printer ! "print"
+      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
+
+      system.stop(firstStep)
+      system.stop(workers)
+      system.stop(stat)
+      system.stop(printer)
+      /*
+!!!!!!PRINT!!!!!!!! nr=16
+ENTRY: maxQueue=2, utilization 15,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347685000,1377347700000,2,15.0,0.0,15.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347700000,1377347715000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347715000,1377347730000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347730000,1377347745000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347745000,1377347760000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347760000,1377347775000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347775000,1377347790000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347790000,1377347805000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347805000,1377347820000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 67,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347820000,1377347835000,1,65.0,0.0,67.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347835000,1377347850000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347850000,1377347865000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347865000,1377347880000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347880000,1377347895000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347895000,1377347910000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 52,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347910000,1377347925000,1,52.0,0.0,52.0,10.0)
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$l],1377347685000,1377347700000,1,0.0,636.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1377347685000,1377347700000,1,0.0,2777.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$e],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1377347685000,1377347700000,1,0.0,2762.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1377347685000,1377347700000,1,0.0,631.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347685000,1377347700000,3,16.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$h],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$b],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1377347685000,1377347700000,1,0.0,629.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1377347685000,1377347700000,1,0.0,2772.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1377347685000,1377347700000,1,0.0,629.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$f],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1377347685000,1377347700000,1,0.0,2760.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$i],1377347685000,1377347700000,1,0.0,657.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$c],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1377347685000,1377347700000,1,0.0,629.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1377347685000,1377347700000,1,0.0,2767.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1377347685000,1377347700000,1,0.0,2760.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$j],1377347685000,1377347700000,1,0.0,646.0,13.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347685000,1377347700000,1,1.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$g],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers],1377347685000,1377347700000,85,6.0,8.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$d],1377347685000,1377347700000,1,0.0,0.0,6.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1377347685000,1377347700000,1,0.0,2788.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347685000,1377347700000,2,15.0,0.0,15.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1377347685000,1377347700000,1,0.0,2760.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$a],1377347685000,1377347700000,1,0.0,0.0,6.0,1007.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$k],1377347685000,1377347700000,1,0.0,641.0,13.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1377347700000,1377347715000,5,0.0,11579.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1377347700000,1377347715000,3,0.0,2355.0,72.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1377347700000,1377347715000,2,0.0,1182.0,50.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1377347700000,1377347715000,6,0.0,5408.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347700000,1377347715000,5,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1377347700000,1377347715000,2,0.0,1519.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1377347700000,1377347715000,2,0.0,1256.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1377347700000,1377347715000,5,0.0,6649.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1377347700000,1377347715000,2,0.0,1263.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1377347700000,1377347715000,2,0.0,1256.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347700000,1377347715000,6,0.0,5401.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1377347700000,1377347715000,6,0.0,10558.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347700000,1377347715000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347700000,1377347715000,6,0.0,7309.0,42.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347700000,1377347715000,1,0.0,17567.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347700000,1377347715000,6,0.0,11314.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1377347700000,1377347715000,2,0.0,1246.0,50.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347700000,1377347715000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1377347700000,1377347715000,2,0.0,1191.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1377347700000,1377347715000,6,0.0,5164.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1377347700000,1377347715000,7,0.0,10114.0,42.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347700000,1377347715000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347700000,1377347715000,1,0.0,17565.0,2.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1377347715000,1377347730000,10,0.0,11656.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1377347715000,1377347730000,1,0.0,578.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347715000,1377347730000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1377347715000,1377347730000,5,0.0,21515.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347715000,1377347730000,6,0.0,24611.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347715000,1377347730000,6,0.0,22909.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347715000,1377347730000,4,0.0,3503.0,75.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347715000,1377347730000,5,0.0,26459.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1377347715000,1377347730000,8,0.0,8837.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347715000,1377347730000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347715000,1377347730000,3,0.0,29988.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347715000,1377347730000,7,0.0,6288.0,87.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347715000,1377347730000,7,0.0,18810.0,40.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347715000,1377347730000,5,0.0,26474.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347715000,1377347730000,9,0.0,7826.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347715000,1377347730000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1377347715000,1377347730000,3,0.0,1161.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1377347715000,1377347730000,8,0.0,5511.0,90.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347715000,1377347730000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347715000,1377347730000,12,0.0,13810.0,97.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1377347730000,1377347745000,9,0.0,8803.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347730000,1377347745000,2,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1377347730000,1377347745000,5,0.0,36426.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347730000,1377347745000,16,0.0,24262.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347730000,1377347745000,11,0.0,34929.0,68.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347730000,1377347745000,11,0.0,25524.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1377347730000,1377347745000,1,0.0,422.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347730000,1377347745000,15,0.0,26202.0,62.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347730000,1377347745000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347730000,1377347745000,6,0.0,37802.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347730000,1377347745000,15,0.0,27444.0,84.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347730000,1377347745000,7,0.0,10134.0,46.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347730000,1377347745000,12,0.0,21940.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347730000,1377347745000,15,0.0,26749.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1377347730000,1377347745000,7,0.0,6308.0,97.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347730000,1377347745000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347730000,1377347745000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1377347730000,1377347745000,3,0.0,44928.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347730000,1377347745000,11,0.0,10826.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347730000,1377347745000,3,0.0,44929.0,15.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347745000,1377347760000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1377347745000,1377347760000,6,0.0,55242.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347745000,1377347760000,18,0.0,21225.0,97.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347745000,1377347760000,3,0.0,59839.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347745000,1377347760000,15,0.0,31188.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1377347745000,1377347760000,15,0.0,31247.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1377347745000,1377347760000,5,0.0,51382.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347745000,1377347760000,19,0.0,26196.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347745000,1377347760000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347745000,1377347760000,16,0.0,37662.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347745000,1377347760000,21,0.0,25395.0,75.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1377347745000,1377347760000,8,0.0,7002.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347745000,1377347760000,17,0.0,27285.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347745000,1377347760000,16,0.0,39952.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347745000,1377347760000,5,0.0,56337.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347745000,1377347760000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347745000,1377347760000,6,0.0,52742.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347745000,1377347760000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1377347745000,1377347760000,10,0.0,47382.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347745000,1377347760000,9,0.0,17067.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347745000,1377347760000,13,0.0,43187.0,66.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347760000,1377347775000,3,0.0,74764.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347760000,1377347775000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1377347760000,1377347775000,16,0.0,54093.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347760000,1377347775000,5,0.0,71250.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347760000,1377347775000,1,0.0,77235.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1377347760000,1377347775000,17,0.0,21473.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347760000,1377347775000,5,0.0,58587.0,17.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347760000,1377347775000,20,0.0,41951.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1377347760000,1377347775000,10,0.0,64833.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347760000,1377347775000,19,0.0,28991.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347760000,1377347775000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347760000,1377347775000,20,0.0,38669.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347760000,1377347775000,19,0.0,21818.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347760000,1377347775000,11,0.0,64306.0,69.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1377347760000,1377347775000,10,0.0,64071.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347760000,1377347775000,18,0.0,29425.0,48.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1377347760000,1377347775000,19,0.0,38698.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347760000,1377347775000,6,0.0,69794.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347760000,1377347775000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347760000,1377347775000,15,0.0,54568.0,64.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347760000,1377347775000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347760000,1377347775000,5,0.0,69293.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1377347760000,1377347775000,13,0.0,18258.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347760000,1377347775000,22,0.0,44880.0,60.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347775000,1377347790000,8,0.0,78770.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347775000,1377347790000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1377347775000,1377347790000,20,0.0,53563.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347775000,1377347790000,5,0.0,79233.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347775000,1377347790000,5,0.0,85150.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347775000,1377347790000,10,0.0,79713.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347775000,1377347790000,5,0.0,75982.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347775000,1377347790000,5,0.0,81160.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347775000,1377347790000,15,0.0,69553.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347775000,1377347790000,1,0.0,92099.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347775000,1377347790000,25,0.0,52385.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1377347775000,1377347790000,5,0.0,86126.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1377347775000,1377347790000,15,0.0,66347.0,37.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347775000,1377347790000,3,0.0,89659.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347775000,1377347790000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347775000,1377347790000,21,0.0,62376.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1377347775000,1377347790000,18,0.0,33954.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347775000,1377347790000,17,0.0,63522.0,40.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347775000,1377347790000,1,0.0,92103.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347775000,1377347790000,20,0.0,28267.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347775000,1377347790000,15,0.0,67308.0,64.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347775000,1377347790000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347775000,1377347790000,25,0.0,56602.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347775000,1377347790000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347775000,1377347790000,6,0.0,84295.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347775000,1377347790000,10,0.0,78763.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1377347775000,1377347790000,30,0.0,44267.0,57.0,1000.0))
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1377347790000,1377347805000,15,0.0,85433.0,47.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1377347790000,1377347805000,3,74.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1377347790000,1377347805000,15,0.0,86451.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1377347790000,1377347805000,10,0.0,93639.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1377347790000,1377347805000,18,0.0,82240.0,52.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1377347790000,1377347805000,5,0.0,96066.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1377347790000,1377347805000,7,0.0,98460.0,9.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1377347790000,1377347805000,11,0.0,91791.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1377347790000,1377347805000,20,0.0,73773.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1377347790000,1377347805000,5,0.0,90846.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1377347790000,1377347805000,25,0.0,56961.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1377347790000,1377347805000,10,0.0,88573.0,51.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1377347790000,1377347805000,19,0.0,48207.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1377347790000,1377347805000,1,8.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1377347790000,1377347805000,26,0.0,62883.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1377347790000,1377347805000,1,0.0,107043.0,2.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1377347790000,1377347805000,25,0.0,69774.0,57.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1377347790000,1377347805000,5,0.0,90850.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1377347790000,1377347805000,13,0.0,89819.0,15.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1377347790000,1377347805000,20,0.0,36544.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1377347790000,1377347805000,20,0.0,72808.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1377347790000,1377347805000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1377347790000,1377347805000,30,0.0,61886.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1377347790000,1377347805000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1377347790000,1377347805000,10,0.0,83435.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1377347790000,1377347805000,15,0.0,85251.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1377347790000,1377347805000,1,0.0,107058.0,2.0,1000.0))
+
+
+      */
+    }
+  }
+}
+
