+++ chapter-perf/src/test/scala/aia/performance/dispatcher/DispatcherThreads2Test.scala
@@ -1,0 +1,84 @@
+package aia.performance.dispatcher
+
+import akka.testkit.TestProbe
+import akka.actor.{Props, ActorSystem}
+import org.scalatest.{WordSpecLike, BeforeAndAfterAll, MustMatchers}
+import akka.routing.RoundRobinPool
+import aia.performance.monitor.ActorStatistics
+import com.typesafe.config.ConfigFactory
+import aia.performance.{SystemMessage, ProcessRequest, PrintMsg}
+import aia.performance.monitor.{MonitorActor, MailboxStatistics, ActorStatistics, MonitorStatisticsActor}
+import concurrent.duration._
+
+class DispatcherThreads2Test extends WordSpecLike
+  with BeforeAndAfterAll
+  with MustMatchers {
+
+  val configuration = ConfigFactory.load("performance/dispatcher")
+  implicit val system = ActorSystem("DispatcherTest", configuration)
+
+  "System" must {
+    "Use multiple threads2" in {
+      //val nrMessages = 15000
+      val nrMessages = 6000
+      val nrWorkers = 100
+      val statDuration = 15000 millis
+
+      val printer = system.actorOf(Props[PrintMsg], "Printer-threads2")
+      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
+        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat-threads2")
+      system.eventStream.subscribe(
+        stat,
+        classOf[ActorStatistics])
+      system.eventStream.subscribe(
+        stat,
+        classOf[MailboxStatistics])
+
+      val end = TestProbe()
+
+      val workers = system.actorOf(
+        RoundRobinPool(nrWorkers).props(
+          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
+            withDispatcher("myMultiThread-dispatcher")),
+        "Workers")
+
+      val firstStep = system.actorOf(Props(new ProcessRequest(10 millis, workers) with MonitorActor).withDispatcher("myMultiThread-dispatcher"), "Entry")
+
+      for (i <- 0 until nrMessages) {
+        firstStep ! new SystemMessage()
+        Thread.sleep(15)
+      }
+      Thread.sleep(10000)
+      printer ! "print"
+      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
+
+      system.stop(firstStep)
+      system.stop(workers)
+      system.stop(stat)
+      system.stop(printer)
+      /*
+!!!!!!PRINT!!!!!!!! nr=15
+ENTRY: maxQueue=3, utilization 29,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822910000,1376822925000,3,29.0,0.0,29.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822925000,1376822940000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822940000,1376822955000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822955000,1376822970000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822970000,1376822985000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822985000,1376823000000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823000000,1376823015000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823015000,1376823030000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823030000,1376823045000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823045000,1376823060000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823060000,1376823075000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823075000,1376823090000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823090000,1376823105000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823105000,1376823120000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823120000,1376823135000,1,66.0,0.0,66.0,10.0)
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$wb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$l],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$e],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376822910000,1376822925000,2,54.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$qb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$h],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ib],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$b],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$fb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$sb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$jb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$f],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Cb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$rb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Gb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Hb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$vb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Eb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$gb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$i],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$kb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$c],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$eb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ob],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ib],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ub],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Db],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$j],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Bb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376822910000,1376822925000,1,25.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$g],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers],1376822910000,1376822925000,100,6.0,19.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$nb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$hb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$zb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$lb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$d],1376822910000,1376822925000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$yb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$xb],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$pb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Jb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822910000,1376822925000,3,29.0,0.0,29.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$tb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1376822910000,1376822925000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Fb],1376822910000,1376822925000,1,0.0,0.0,24.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ab],1376822910000,1376822925000,1,0.0,0.0,25.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$a],1376822910000,1376822925000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1376822910000,1376822925000,1,0.0,0.0,27.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$k],1376822910000,1376822925000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$mb],1376822910000,1376822925000,1,0.0,0.0,26.0,1000.0))
+
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$G],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$wb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$l],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$r],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$e],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$3],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$u],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$z],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$m],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376823105000,1376823120000,9,132.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$W],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$qb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$5],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$P],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$h],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ib],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$b],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$8],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$J],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$n],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$2],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$bb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$fb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$sb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$s],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$o],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$jb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$f],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$w],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$4],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Cb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$6],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$rb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Gb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$L],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Hb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$vb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Eb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$gb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Y],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$~],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Q],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$i],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$kb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$c],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$eb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ob],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$K],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$A],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ib],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$9],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ub],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$U],1376823105000,1376823120000,1,0.0,0.0,66.0,1001.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Db],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$p],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$t],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$+],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$x],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$C],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$M],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$j],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$E],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Bb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376823105000,1376823120000,1,66.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$R],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$g],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$cb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$O],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$B],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$0],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$nb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$ab],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Z],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$hb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$zb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$H],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$N],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$lb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$d],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$X],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$F],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$yb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$xb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$q],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$pb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Jb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376823105000,1376823120000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$V],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$tb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$v],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$y],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$D],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1376823105000,1376823120000,1,0.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Fb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$Ab],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$T],1376823105000,1376823120000,1,0.0,0.0,66.0,1003.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$a],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$7],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$1],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$I],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$db],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$S],1376823105000,1376823120000,1,0.0,0.0,66.0,1004.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$k],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Workers/$mb],1376823105000,1376823120000,1,0.0,0.0,66.0,1000.0))
+       */
+
+    }
+  }
+}
+
