+++ chapter-perf/src/test/scala/aia/performance/dispatcher/DispatcherThreadsTest.scala
@@ -1,0 +1,87 @@
+package aia.performance.dispatcher
+
+import akka.testkit.TestProbe
+import akka.actor.{Props, ActorSystem}
+import org.scalatest.{WordSpecLike, BeforeAndAfterAll, MustMatchers}
+import akka.routing.RoundRobinPool
+import aia.performance.monitor.ActorStatistics
+import com.typesafe.config.ConfigFactory
+import aia.performance.{SystemMessage, ProcessRequest, PrintMsg}
+import aia.performance.monitor.{MonitorActor, MailboxStatistics, ActorStatistics, MonitorStatisticsActor}
+import concurrent.duration._
+
+class DispatcherThreadsTest extends WordSpecLike
+  with BeforeAndAfterAll
+  with MustMatchers {
+
+  val configuration = ConfigFactory.load("performance/dispatcher")
+  implicit val system = ActorSystem("DispatcherTest", configuration)
+
+  "System" must {
+    "Use multiple threads" in {
+      //val nrMessages = 15000
+      val nrMessages = 6000
+      val nrWorkers = 100
+      val statDuration = 15000 millis
+
+      val printer = system.actorOf(Props[PrintMsg], "Printer-threads")
+      val stat = system.actorOf(Props(new MonitorStatisticsActor(period = statDuration, processMargin = 1000,
+        storeSummaries = printer)).withDispatcher("my-pinned-dispatcher"), "stat-threads")
+      system.eventStream.subscribe(
+        stat,
+        classOf[ActorStatistics])
+      system.eventStream.subscribe(
+        stat,
+        classOf[MailboxStatistics])
+
+      val end = TestProbe()
+
+      val workers = system.actorOf(
+        RoundRobinPool(nrWorkers).props(
+          Props(new ProcessRequest(1 second, end.ref) with MonitorActor).
+            withDispatcher("myMultiThread-dispatcher")),
+        "Workers-threads")
+
+      val firstStep = system.actorOf(
+        Props(new ProcessRequest(10 millis, workers) with MonitorActor).
+          withDispatcher("my-pinned-dispatcher"), "Entry")
+
+      for (i <- 0 until nrMessages) {
+        firstStep ! new SystemMessage()
+        Thread.sleep(15)
+      }
+      Thread.sleep(10000)
+      printer ! "print"
+      val msg = end.receiveN(n = nrMessages, max = 30 seconds)
+
+      system.stop(firstStep)
+      system.stop(workers)
+      system.stop(stat)
+      system.stop(printer)
+      /*
+!!!!!!PRINT!!!!!!!! nr=15
+ENTRY: maxQueue=2, utilization 34,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376821995000,1376822010000,2,34.0,0.0,34.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822010000,1376822025000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822025000,1376822040000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822040000,1376822055000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822055000,1376822070000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822070000,1376822085000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822085000,1376822100000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822100000,1376822115000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822115000,1376822130000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822130000,1376822145000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822145000,1376822160000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822160000,1376822175000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822175000,1376822190000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822190000,1376822205000,1,66.0,0.0,66.0,10.0)
+ENTRY: maxQueue=1, utilization 66,000000:  StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822205000,1376822220000,1,66.0,0.0,66.0,10.0)
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/$A],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ib],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$V],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$jb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376821995000,1376822010000,3,65.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$8],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$N],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$F],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$xb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$g],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Jb],1376821995000,1376822010000,1,0.0,0.0,28.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$c],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$cb],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$4],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$9],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$y],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$D],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$0],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$s],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$O],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$w],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$K],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ob],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ab],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$nb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$vb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$q],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Db],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$5],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$m],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$i],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$U],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$E],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$S],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$lb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$eb],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$qb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$tb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$hb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Fb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$z],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$~],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$mb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$B],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$2],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$h],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$d],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Eb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$n],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$j],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$H],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$t],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$x],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$L],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$P],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$X],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$db],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$1],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$wb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$pb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$yb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Hb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$k],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376821995000,1376822010000,1,30.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$6],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$o],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$sb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$gb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$M],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$I],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$C],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Y],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$a],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$u],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ib],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$+],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Q],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Gb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$G],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$W],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$e],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$rb],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$b],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ub],1376821995000,1376822010000,1,0.0,0.0,30.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376821995000,1376822010000,2,34.0,0.0,34.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$R],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ab],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$J],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Z],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$3],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Cb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$kb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$v],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$7],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$r],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$l],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$p],1376821995000,1376822010000,1,0.0,0.0,34.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Bb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$fb],1376821995000,1376822010000,1,0.0,0.0,31.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$T],1376821995000,1376822010000,1,0.0,0.0,33.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$zb],1376821995000,1376822010000,1,0.0,0.0,29.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$bb],1376821995000,1376822010000,1,0.0,0.0,32.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$f],1376821995000,1376822010000,1,0.0,0.0,35.0,1000.0))
+
+Received List(StatisticsSummary(Actor[akka://DispatcherTest/user/$A],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ib],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$V],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$jb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/stat],1376822205000,1376822220000,8,132.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$8],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$N],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$F],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$xb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$g],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Jb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$c],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$cb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$4],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$9],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$y],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$D],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$0],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$s],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$O],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$w],1376822205000,1376822220000,1,0.0,0.0,66.0,1002.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$K],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ob],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ab],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$nb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$vb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$q],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Db],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$5],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$m],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$i],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$U],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$E],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$S],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$lb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$eb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$qb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$tb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$hb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Fb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$z],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$~],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$mb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$B],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$2],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$h],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$d],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Eb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$n],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$j],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$H],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$t],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$x],1376822205000,1376822220000,1,0.0,0.0,66.0,1001.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$L],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$P],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$X],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$db],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$1],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$wb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$pb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$yb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Hb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$k],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/system/testActor1],1376822205000,1376822220000,1,66.0,0.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$6],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$o],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$sb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$gb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$M],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$I],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$C],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Y],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$a],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$u],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Ib],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$+],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Q],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Gb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$G],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$W],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$e],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$rb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$b],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ub],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Entry],1376822205000,1376822220000,1,66.0,0.0,66.0,10.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$R],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$ab],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$J],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Z],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$3],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/Printer],1376822205000,1376822220000,1,0.0,1.0,0.0,0.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Cb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$kb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$v],1376822205000,1376822220000,1,0.0,0.0,66.0,1004.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$7],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$r],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$l],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$p],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$Bb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$fb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$T],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$zb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$bb],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0), StatisticsSummary(Actor[akka://DispatcherTest/user/$f],1376822205000,1376822220000,1,0.0,0.0,66.0,1000.0))
+
+       */
+
+    }
+  }
+}
+
